+++ conf.lua
@@ -1,8 +1,6 @@
 function love.conf(t)
-	t.console = true                  -- Attach a console (boolean, Windows only)
-
-	t.window.title = "Philtre2D Test Project"        -- The window title (string)
-	t.window.width = 800               -- The window width (number)
-	t.window.height = 600              -- The window height (number)
+	t.window.title = "Philtre2D"
+	t.window.width = 800
+	t.window.height = 600
 	t.window.resizable = true
 end