+++ history.lua
@@ -1,91 +1,121 @@
 -- Manage a command history with undo, redo, etc.
 
-local function register(history, name, perform, revert, update)
-	history.functions[name] = {
+-- local references to library functions
+local append = table.insert
+local removeLast = table.remove
+
+local function appendAll(seq, ...)
+	for i=1,select('#', ...) do
+		append(seq, select(i, ...))
+	end
+end
+
+local function selector(self, fn)
+	self.selectors[fn] = fn
+end
+
+local function expand(self, args)
+	if type(args) ~= 'table' then return args end
+	local out = {}
+	for i,arg in ipairs(args) do
+		if type(arg) == 'table' and self.selectors[arg[1]] then
+			local fn, fnArgs = arg[1], {unpack(arg, 2)}
+			appendAll(out, fn(expand(self, fnArgs)))
+		else
+			append(out, arg)
+		end
+	end
+	return unpack(out)
+end
+
+local function command(self, name, perform, revert, update)
+	self.commands[name] = {
 		perform = perform,
 		revert = revert,
 		update = update
 	}
 end
 
-local function newCmd(fns, name, args)
+local function commandInstance(self, fns, name, args)
 	return {
 		future = {},
 		perform = fns.perform,  revert = fns.revert,
 		update = fns.update,
-		name = name,  args = args,
-		undoArgs = { fns.perform(unpack(args)) }
+		name = name,  args = {unpack(args)},
+		undoArgs = { fns.perform(expand(self, args)) }
 	}
 end
 
-local function perform(history, name, ...)
-	local fns = history.functions[name]
-	local cmd = newCmd(fns, name, {...})
-	local prev = history.mostRecent
+local function perform(self, name, ...)
+	local fns = self.commands[name]
+	local cmd = commandInstance(self, fns, name, {...})
+	local prev = self.mostRecent
 	cmd.past = prev
-	table.insert(prev.future, 1, cmd)
-	history.mostRecent = cmd
+	append(prev.future, 1, cmd)
+	self.mostRecent = cmd
 	-- Return undoArgs for interactive commands, e.g. when
 	-- interactively creating an object you may want to move or
 	-- resize it immediately.
 	return cmd.undoArgs
 end
 
-local function undo(history)
-	local cmd = history.mostRecent
-	cmd.revert(unpack(cmd.undoArgs))
-	history.mostRecent = cmd.past
+local function undo(self)
+	local cmd = self.mostRecent
+	cmd.revert(expand(self, cmd.undoArgs))
+	self.mostRecent = cmd.past
 end
 
-local function redo(history)
-	local cmd = history.mostRecent.future[1]
+local function redo(self)
+	local cmd = self.mostRecent.future[1]
 	if cmd then
-		cmd.undoArgs = { cmd.perform(unpack(cmd.args)) }
-		history.mostRecent = cmd
+		cmd.undoArgs = { cmd.perform(expand(self, cmd.args)) }
+		self.mostRecent = cmd
 	end
 end
 
-local function update(history, ...)
-	local cmd = history.mostRecent
+local function update(self, ...)
+	local cmd = self.mostRecent
 	if cmd.update then
-		cmd.args = cmd.update(...)
+		cmd.args = {cmd.update(...)}
 	else
+		-- TODO - this is wrong: perform takes a name.
+		-- Should we give an error?  Or...?
 		cmd.args = {...}
 		cmd.perform(...)
 	end
 end
 
-local function cancel(history)
-	local cmd = history.mostRecent
+local function cancel(self)
+	local cmd = self.mostRecent
 	if cmd.past ~= cmd then
-		cmd.revert(unpack(cmd.undoArgs))
+		cmd.revert(expand(self, cmd.undoArgs))
 		for _,v in ipairs(cmd.future) do
-			table.insert(cmd.past.future, v)
+			append(cmd.past.future, v)
 		end
 		local futures = cmd.past.future
 		for i=1,#futures do
 			if futures[i] == cmd then
-				table.remove(futures, i)
+				removeLast(futures, i)
 				break
 			end
 		end
-		history.mostRecent = cmd.past
+		self.mostRecent = cmd.past
 	end
 end
 
-local function chooseFuture(history, n)
-	local f = history.mostRecent.future
+local function chooseFuture(self, n)
+	local f = self.mostRecent.future
 	if n <= #f and n >= 1 then
-		table.insert(f, 1, table.remove(f, n))
+		append(f, 1, removeLast(f, n))
 	end
 end
 
-local function futureCount(history)
-	return #history.mostRecent.future
+local function futureCount(self)
+	return #self.mostRecent.future
 end
 
 local methods = {
-	register = register,
+	command = command,  selector = selector,
 	perform = perform,  undo = undo,  redo = redo,
 	update = update,  cancel = cancel,
 	chooseFuture = chooseFuture,  futureCount = futureCount
@@ -95,11 +125,11 @@ local class = { __index = methods }
 local function noop() end
 
 local function new()
+	local h = { commands = {}, selectors = {} }
+	-- Ensure that there's always a dummy command in the history
+	-- to make the code simpler.
 	local noops = { perform = noop,  revert = noop }
-	local h = {
-		functions = {},
-		mostRecent = newCmd(noops, nil, {})
-	}
+	h.mostRecent = commandInstance(h, noops, nil, {})
 	h.mostRecent.past = h.mostRecent
 	return setmetatable(h, class)
 end