+++ main.lua
@@ -1,183 +1,174 @@
--- Example thingy
-
 require('engine.all')
+local Bezier = require('engine.bezier')
 
-local Box = require('box')
-local flux = require('lib.flux')
+local History = require('history')
+
+
+----------------------------------------------------------------
+-- Vector math on arrays of arbitrary size.
+
+local function lerp(t, a, b)
+	local out = {}
+	for i=1,#a do out[i] = a[i] + t * (b[i] - a[i]) end
+	return out
+end
+
+local function distanceSquared(a, b)
+	local d2 = 0
+	for i=1,#a do
+		local dx = b[i] - a[i]
+		d2 = d2 + dx*dx
+	end
+	return d2
+end
+
+
+----------------------------------------------------------------
+-- Curve commands.
+
+local function extendBezier(curve, x, y)
+	local endpoint = {x, y}
+	if #curve > 0 then
+		local beginpoint = curve[#curve]
+		table.insert(curve, lerp(1/3, beginpoint, endpoint))
+		table.insert(curve, lerp(2/3, beginpoint, endpoint))
+	end
+	table.insert(curve, endpoint)
+	return curve
+end
+
+local function retractBezier(curve)
+	local cp1 = table.remove(curve)
+	local cp2 = table.remove(curve)
+	local ep = table.remove(curve)
+	return curve, cp1, cp2, ep
+end
+
+local function extendBezierWith(curve, cp1, cp2, ep)
+	table.insert(curve, cp1)
+	table.insert(curve, cp2)
+	table.insert(curve, ep)
+end
+
+local function moveBezierPoint(curve, n, x, y)
+	local p = curve[n]
+	local undoX, undoY = p[1], p[2]
+	p[1], p[2] = x, y
+	return curve, n, undoX, undoY
+end
+
+local function nearestPoint(curve, x, y)
+	local q = {x, y}
+	local nearest, dist2 = false, math.huge
+	for i,p in ipairs(curve) do
+		local d2 = distanceSquared(p, q)
+		if d2 < dist2 then dist2, nearest = d2, i end
+	end
+	return nearest, math.sqrt(dist2)
+end
+
+
+----------------------------------------------------------------
+-- Main love callbacks.
 
 function love.load()
-	-- init physics world - disable pre- and post-solve callbacks
-	local phys_obj = physics.new(0, 500, true, false, false, true, true)
-	world = phys_obj.world
-
-	local red = { 180, 23, 20 }
-	local green = { 30, 200, 25 }
-
-	draw_order = DrawOrder.new('default')
-	draw_order:add_layer(1, "bg")
-
-	img_yellow_blob = love.graphics.newImage('yellow-blob.png')
-	local img_sq_64 = love.graphics.newImage('square_64.png')
-	local img_rect_128x256 = love.graphics.newImage('rect_128x256.png')
-
-	local fnt_opensans_reg = 'assets/fonts/OpenSans-Regular.ttf'
-
-	scene = T.new(draw_order, {
-		mod(Box.new(300, 200, 30, 30, red), {
-			name = 'red-box',
-			children = {
-				mod(Box.new(20, -25, 8, 8, red), {name = "first red child", angle = math.pi/6}),
-				mod(Box.new(20, 25, 18, 4, red), {name = "red arm", angle = math.pi}),
-				mod(Box.new(-20, 0, 18, 4, red), {name= "red spinny thing", angle = math.pi, vAngle = -3}),
-				mod(Body.new(world, 'kinematic', -20, -10, {{'circle', {25}}}),
-					{children = {Sprite.new(img_yellow_blob, 'center', 'center', 10, 0, 0, 0.25)}
-				})
-			},
-			v = { x = 60, y = 20 },
-			vAngle = math.pi,
-			sx = 1
-		}),
-		mod(Box.new(250, 100, 30, 30, green), {
-			angle = 0.5,
-			name = 'green-box',
-			layer = 'bg',
-			children = {
-				mod(Box.new(8, -25, 8, 8, green), {
-					angle=-math.pi/6,
-					script = {
-						init = function(self)
-							self.t, self.x0, self.y0 = 0, self.pos.x, self.pos.y
-						end,
-						update = function(self, dt)
-							self.t = (self.t + 10*dt) % (2*math.pi)
-							self.pos.x = self.x0 - 5*math.sin(self.t)
-							self.pos.y = self.y0 + 5*math.cos(self.t)
-						end
-					}
-				}),
-				Text.new(0, 20, 0, 'Text Object')
-			}
-		}),
-		mod(Sprite.new(img_yellow_blob, 'center', 'center', 100, 100), {
-			update = function(self, dt)
-				self.angle = self.angle - dt * 4*math.pi
-			end
-		}),
-		mod(Body.new(world, 'dynamic', 400, 100, { {'circle', {50}, restitution = 0.2}, { 'circle', {-34, 0, 20}}}, {gScale=5}), {
-			name = 'physics body',
-			children = {
-				mod(Sprite.new(img_yellow_blob, 'center', 'center', 15, 0, math.pi/6), {name='physics body child'})
-			},
-		}),
-		mod(Body.new(world, 'static', 400, 550, { {'rectangle', {600, 50}} }), {name = "ground"}),
-		Camera.new(300, 300, 0, 0.8),
-		phys_obj
-	})
-
-	if scene:get('/red-box') ~= scene.children[1] then
-		print('/red-box:', scene:get('/red-box'), scene.children[1])
-		error('Getting red box')
-	end
-
-	if scene:get('/green-box') ~= scene.children[2] then
-		print('/green-box:', scene:get('/green-box'), scene.children[1])
-		error('Getting green box')
-	end
-
-	-- Input test.
-	Input.init()
-	Input.bind({
-		{'quit', 'button', 'key', 'x'},
-		{'x', 'axis_from_buttons', 'scancode', 'f', 'scancode', 's'},
-		{'y', 'axis_from_buttons', 'scancode', 'd', 'scancode', 'e'},
-		{'quit', 'button_from_axis', 'joystick1', 'triggerright'},
-		{'x', 'axis', 'joystick1', 'leftx'},
-		{'y', 'axis', 'joystick1', 'lefty'},
-	})
-	-- Move green box.
-	local green_box = scene:get('/green-box')
-	green_box.input = function(self, name, value, change)
-		local speed = 100
-		if name == 'x' then 
-			self.v.x = value * speed
-		elseif name == 'y' then
-			self.v.y = value * speed
-		end
-	end
-	Input.enable(green_box)
-	-- Quit with x and right trigger (not escape).
-	global_input = {
-		input = function(self, name, value, change)
-			if name == 'quit' then
-				love.event.quit()
-			end
-		end
+	pickDistance = 10
+	curve = {
+		highlight = false,
+		{100, 100}, {133, 133}, {167, 167}, {200, 200}
 	}
-	Input.enable(global_input)
+	dragging = false
+	edits = History.new()
+	edits:command('extendBezier', extendBezier, retractBezier)
+	edits:command('retractBezier', retractBezier, extendBezierWith)
+	edits:command('moveBezierPoint', moveBezierPoint, moveBezierPoint, moveBezierPoint)
 
-	flux.to(scene.children[1], 2, {sx=2}):oncomplete(function()
-		flux.to(scene.children[1], 2, {sx=1})
-	end)
-
-	gui_root_obj = Gui_root.new()
-
-	gui_scene = T.new(nil, {
-		mod(gui_root_obj, {
-			children = {
-				mod(Gui_sprite.new(img_rect_128x256, 'right', 'top', -10, 10, 0, 1.5, 1.5, 'right', 'top'), {
-					color = {255, 255, 255, 127},
-					children = {
-						Gui_sprite.new(img_sq_64, 'right', 'center', 0, 0, 0, 0.8, 0.8, 'right', 'bottom'),
-						Gui_text.new(0, 10, 0, 'Gui Text Object', fnt_opensans_reg, 16, 300, 'center', 1/1.5, 1/1.5, 'center', 'top')
-					}
-				}),
-				mod(Gui_sprite.new(img_sq_64, 'center', 'bottom', 0, -10, 0, 780/64, 0.5, 'center', 'bottom', 'stretch'), {
-					color = {255, 255, 255, 127}
-				})
-			}
-		})
-	})
-end
-
-function love.resize(w, h)
-	Camera.window_resized(w, h)
-	gui_root_obj:window_resized(w, h)
-end
-
-function love.update(dt)
-	flux.update(dt)
-	scene:update(dt)
-	gui_scene:update(dt)
-	Camera.update(dt)
+	love.graphics.setLineWidth(3)
 end
 
 function love.draw()
-	Camera.apply_transform()
-	scene:draw()
-	Camera.reset_transform()
-	gui_scene:draw()
-end
+	if #curve >= 4 then
+		local points = {}
+		for i=1,#curve-3,3 do
+			if i > 1 then table.remove(points) end
+			local b = {unpack(curve, i, i+3)}
+			Bezier.toPolyline(b, 0.5, points)
+		end
+		local coords = {}
+		for _,p in ipairs(points) do
+			table.insert(coords, p[1])
+			table.insert(coords, p[2])
+		end
+		love.graphics.line(coords)
+	end
 
-function love.keypressed(k, s)
-	if k == 'escape' then
-		love.event.quit()
+	local lw = love.graphics.getLineWidth()
+	local r = math.ceil(1.5 * lw)
+	local q = false
+	for i,p in ipairs(curve) do
+		local x, y = unpack(p)
+		local h = (i == curve.highlight) and 1.5 or 1
+		love.graphics.circle('fill', x, y, r*h)
+		if q and i % 3 ~= 0 then
+			local w = love.graphics.getLineWidth()
+			local c = {love.graphics.getColor()}
+			local a = c[4]
+			c[4] = c[4] / 2
+			love.graphics.setLineWidth(1)
+			love.graphics.setColor(c)
+			love.graphics.line(q[1], q[2], x, y)
+			c[4] = a
+			love.graphics.setLineWidth(w)
+			love.graphics.setColor(c)
+		end
+		q = p
 	end
 end
 
-function love.mousepressed(x, y, button, isTouch)
-	if button == 1 then
-		x, y = Camera.screen_to_world(x, y)
-		scene:add(mod(Sprite.new(img_yellow_blob, 'center', 'center', x, y), {name='bob'}))
-	elseif button == 2 then
-		local o
-		for k, v in pairs(scene.children) do
-			if v then o = v break end
-		end
-		if o then
-			print("removing: ", o.path)
-			scene:remove(o)
+function love.mousemoved(x, y)
+	if dragging then
+		edits:update(curve, dragging, x, y)
+	else
+		local i, d = nearestPoint(curve, x, y)
+		if i and d <= pickDistance then
+			curve.highlight = i
 		else
-			print("Object to remove not found")
+			curve.highlight = false
 		end
+	end
+end
+
+function love.mousepressed(x, y, b)
+	local n, d = nearestPoint(curve, x, y)
+	if d > pickDistance then n = false end
+	if b == 1 then
+		if n then
+			dragging = n
+			edits:perform('moveBezierPoint', curve, n, x, y)
+		else
+			edits:perform('extendBezier', curve, x, y)
+		end
+	elseif b == 2 then
+		edits:perform('retractBezier', curve)
+	end
+end
+
+function love.mousereleased(x, y, b)
+	if b == 1 and dragging then
+		dragging = false
+	end
+end
+
+function love.keypressed(k, s)
+	local ctrl = love.keyboard.isDown('lctrl', 'rctrl')
+	local shift = love.keyboard.isDown('lshift', 'rshift')
+	if ctrl then
+		if k == 'z' then
+			if(shift) then edits:redo() else edits:undo() end
+		elseif k == 'y' then
+			edits:redo()
+		end
+	else
+		if k == 'escape' then love.event.quit() end
 	end
 end