+++ doc/editor/history.md
@@ -4,14 +4,19 @@ History
 A history is a list of commands which have been performed and a
 dictionary of the command types that it knows how to do/undo.
 
-You can `register` a command, giving it a name and the following
-functions:
+	local History = require 'history'
+	local h = History.new()
 
-* `perform(performArgs) -> undoArgs`
+Start by registering commands, associating their names with
+the appropriate functions.
 
-* `revert(undoArgs)`
+* `h:command(name, perform, revert, update)`
 
-* `update(...) -> performArgs`
+	* `perform(performArgs) -> undoArgs`
+
+	* `revert(undoArgs)`
+
+	* `update(...) -> performArgs`
 
 The `perform` function is used both to perform the command in
 the first place and to redo it.  The `revert` function is used
@@ -20,7 +25,9 @@ implementing interactive commands (move, resize, etc.) where you
 want to modify the arguments of the command that's already in
 the history instead of adding a million tiny move commands.
 
-Once you have regsistered your commands, you can use the history
+----
+
+Once you have registered your commands, you can use the history
 methods:
 
 * `h:perform(name, args...) -> undoArgs` - Perform a command and
@@ -37,6 +44,8 @@ methods:
 * `h:cancel()` - Cancels the most recent command, removing it
   completely.
 
+----
+
 Note that we do not forget future commands when you perform a
 new one, we simply shift them to an alternate future.  So you
 can ask how many futures there are with `futureCount` and choose
@@ -45,3 +54,37 @@ one with `chooseFuture`.
 * `h:futureCount() -> number`
 
 * `h:chooseFuture(number)`
+
+----
+
+If you undo/redo the creation of an object, it will most likely
+create a different object.  If the argument lists for subsequent
+commands refer directly to the old object (which doesn't exist
+in the world any more) then they will no longer do anything
+useful.
+
+You *could* come up with some scheme to refer to objects
+indirectly, and ensure that if you undo/redo creation, it always
+creates an object with the same reference.  Or you could
+register selection functions and have your argument lists invoke
+those to select the object(s) to operate on.
+
+* `h:selector(fn)` - Register `fn` as a selector function.  Now
+  when `fn` is encountered in argument lists, the function will
+  be invoked on the following argument, and its return value(s)
+  will be used.
+
+
+```lua
+-- pickObject(x, y) -> object under point.
+h:selector(pickObject)
+
+-- selection(n) -> nth most recent selection.
+h:selector(selection)
+
+-- Get object as if you clicked at (100, 230) and make it 50% bigger.
+h:perform('resize',  pickObject, {100,230},  1.5)
+
+-- Get most recently selection and make it 25% smaller.
+h:perform('resize',  selection, 1,  0.75)
+```